move log viewer to new components dir
