+++ src/components/LogViewer.svelte
